Added Optcode::close
See this exampe Websocket packet i received from a Websocket Server

0x881103E86E6F726D616C6C7920636C6F736564 -> ��normally closed 
0x88 		                                                        -> finish = true & optcode = close
   0x11 	                                                                -> 17 Byte Payload Length
      0x03E8                                                               -> 1000 = Websocket Close Code
             0x6E6F726D616C6C7920636C6F736564    -> normally closed
looks like a correct packet
the ignoring of optcode::close result in returning processedBytes = 2
in my case this results in a loop that parses the given Packet several times until the processesBytes > packet.size()

Please see this changes. I'm not sure if this is the best solution to solve this problem.